feat(client): tracks search and results refacto
